requirements.txt
# Azure Functions
azure-functions>=1.18.0

# Azure Services
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0

# Microsoft Agent Framework
azure-ai-projects>=1.0.0
openai>=1.12.0
azure-core>=1.29.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23

# Vector Store & Embeddings
faiss-cpu>=1.7.4
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
